backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.1.0
gunicorn==21.2.0
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
